requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.5
django-cloudinary-storage==0.3.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
mutagen==1.45.1
Pillow==9.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
